버그 수정 수치 조정

+++ A07_Roguelike/Assets/Scripts/AttackData/RangedAttackController.cs
@@ -54,7 +54,7 @@ public class RangedAttackController : MonoBehaviour
             HealthController healthSystem = collision.GetComponent<HealthController>();
             if(healthSystem != null) 
             {
-                healthSystem.ChangeHealth(-_attackData.power);
+                healthSystem.ChangeHealth(-(int)_attackData.power);
                 if (_attackData.isOnKnockback)
                 {
                     TopDownMovement movement = collision.GetComponent<TopDownMovement>();
